refactor: move parser/ into cmtraceopen-parser crate; split parse_file
Completes Phase 1 core extraction. Every log-format parser module
(burn, cbs, ccm, cmtlog, panther, detect, dhcp, dism, dns_debug,
dns_types, iis_w3c, intune_macos, msi, patchmypc_detection, plain,
psadt, registry, reporting_events, secureboot_log, severity, simple,
timestamped) moves into crates/cmtraceopen-parser/src/parser/.

Surfaced dependency: parser/ccm.rs calls intune::ime_parser for the
IME CCM specialization. ime_parser.rs is itself pure (no native
deps), so it moves with the parser into crates/cmtraceopen-parser/
src/intune/ime_parser.rs. src-tauri/src/intune/mod.rs re-exports it
via `pub use cmtraceopen_parser::intune::ime_parser;` so call sites
referencing crate::intune::ime_parser::* keep resolving.

Stays in src-tauri:
- parser/dns_audit.rs — feature-gated EVTX parser, native-only.
- fs-based parse_file + read_file_content — moved into a thin
  src-tauri/src/parser/mod.rs shim that re-exports crate items and
  adds the native-only concerns (ETL rejection, DNS EVTX routing,
  std::fs::read with BOM + Windows-1252 fallback).

New pure entry point in the crate:
  parser::parse_content(content, file_path, file_size)
    -> (ParseResult, ResolvedParser)
Replaces the old `parse_file`'s post-read logic. Platform-agnostic
and wasm-compatible.

Verified:
  cargo check -p cmtraceopen-parser                         0.77s
  cargo build -p cmtraceopen-parser --target wasm32-unknown-unknown
                                                            7.78s
  cargo test -p cmtraceopen-parser --lib                    227 passed

src-tauri full-workspace verification still blocked by the ARM64
clang prereq for the `ring` crate in this shell — to be verified
in the user's usual dev environment (VS BuildTools + LLVM).
