+++ crates/cmtraceopen-parser/src/lib.rs
@@ -8,4 +8,6 @@
 // No Tauri, no tokio, no notify, no evtx, no windows/winreg, no rayon, no filesystem I/O.
 
 pub mod error_db;
+pub mod intune;
 pub mod models;
+pub mod parser;
