+++ crates/cmtraceopen-parser/src/parser/burn.rs
