+++ crates/cmtraceopen-parser/src/parser/cbs.rs
