+++ crates/cmtraceopen-parser/src/parser/ccm.rs
