+++ crates/cmtraceopen-parser/src/parser/detect.rs
