+++ crates/cmtraceopen-parser/src/parser/dns_types.rs
