+++ crates/cmtraceopen-parser/src/parser/intune_macos.rs
