+++ crates/cmtraceopen-parser/src/parser/msi.rs
