+++ crates/cmtraceopen-parser/src/parser/panther.rs
