+++ crates/cmtraceopen-parser/src/parser/patchmypc_detection.rs
