+++ crates/cmtraceopen-parser/src/parser/plain.rs
