+++ crates/cmtraceopen-parser/src/parser/psadt.rs
