+++ crates/cmtraceopen-parser/src/parser/registry.rs
