+++ crates/cmtraceopen-parser/src/parser/reporting_events.rs
