+++ crates/cmtraceopen-parser/src/parser/secureboot_log.rs
