+++ crates/cmtraceopen-parser/src/parser/severity.rs
