+++ crates/cmtraceopen-parser/src/parser/simple.rs
