+++ crates/cmtraceopen-parser/src/parser/timestamped.rs
